+++ cmake/SuperBuild.cmake
@@ -1,5 +1,5 @@
 ################################################################################
-# Copyright (C) 2015-2019 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+# Copyright (C) 2015-2023 Commissariat a l'energie atomique et aux energies alternatives (CEA)
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -22,14 +22,10 @@
 # THE SOFTWARE.
 ################################################################################
 
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.10...3.25)
 
 include(GNUInstallDirs)
 include(ExternalProject)
-
-if(POLICY CMP0114)
-	cmake_policy(SET CMP0114 OLD)
-endif()
 
 ### Generate a build command to build a subproject with access to its dependencies
 #