+++ plugins/decl_netcdf/cmake/FindAstyle.cmake
@@ -25,7 +25,7 @@ Astyle::astyle
 
 #]=======================================================================]
 
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.10...3.25)
 list(INSERT CMAKE_MODULE_PATH 0 "${CMAKE_CURRENT_LIST_DIR}")