+++ plugins/decl_netcdf/cmake/FindNetCDF.cmake
@@ -63,7 +63,7 @@ includes:
 
 #]==]
 
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.10...3.25)
 
 
 set(_NetCDF_features_list CDF5 DAP DAP2 DAP4 DISKLESS HDF4 HDF5 JNA MMAP NC2 NC4 PARALLEL PARALLEL4 PNETCDF)