+++ tools/pdicfg_validator/cmake/FindPyYAML.cmake
@@ -23,7 +23,7 @@
 # THE SOFTWARE.
 ################################################################################
 
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.10...3.25)
 
 include(FindPackageHandleStandardArgs)