+++ CHANGELOG.md
@@ -20,7 +20,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 * Updated the embedded version of PyYAML to 6.0.1 to fix an incompatibility with
   cython 3, fixes
-  [#437](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/437)
+  [#439](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/439)
+* Update PDI's supported versions of CMake from 3.10 to 3.25.
+  Fixed CMake policy warnings introduced in v3.24 and v3.27.6, closes [#437](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/437)
 
 
 ### Security